Moving fetcher test helpers to separate file
